+++ src/model/Model_Payee.h
@@ -60,9 +60,9 @@ public:
 
     bool remove(int64 id);
 
-    const std::map<wxString, int> all_payees(bool excludeHidden = false);
+    const std::map<wxString, int64> all_payees(bool excludeHidden = false);
     const wxArrayString all_payee_names();
-    const std::map<wxString, int> used_payee();
+    const std::map<wxString, int64> used_payee();
 
     static bool is_hidden(int64 id);
     static bool is_hidden(const Data* record);
